fix(opencv): correct gocv vision-tag build + real E2E proof (helix HXC-943)
The //go:build vision path never compiled before (no OpenCV installed), hiding
two real defects surfaced once built against OpenCV 4.13 + gocv v0.43:
  - detector_vision.go: gocv.IMReadGrayscale -> IMReadGrayScale (4 sites)
  - differ_vision.go: SSIM return float32 maxVal -> float64(maxVal)

Add real_vision_e2e_test.go (//go:build vision): synthetic PNGs pushed through
real gocv (Canny/contours, template match, SSIM, change-mask, dominant colors)
with asserted detection results — sink-side proof the feature works, per
HelixConstitution CLAUDE-1.

+++ pkg/opencv/differ_vision.go
@@ -49,7 +49,7 @@ func (d *realDiffer) SSIM(img1, img2 []byte) (float64, error) {
 	gocv.MatchTemplate(gray1, gray2, &result, gocv.TmCcoeffNormed, mask)
 
 	_, maxVal, _, _ := gocv.MinMaxLoc(result)
-	return maxVal, nil
+	return float64(maxVal), nil
 }
 
 // PixelDiff computes a pixel-level diff image between two images.
